requirements.txt
black==22.3.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
docstring-to-markdown==0.15
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
ujson==5.10.0
Werkzeug==3.0.4
